LED blink code
This is a code that requires that a led is connected to analog pin 0 and the "switch" is connected to digital pin 12. When that is true, the led will alternate between on and off every quarter second when the "switch" is powered, and every 1.5 seconds when the "switch" is unpowered.